refactor(listener): separated relay listener into files
